requirements.txt
numpy<2
Flask==3.0.3
torch==1.10.0+cpu
torchvision==0.11.1+cpu
Pillow==8.3.1
tensorflow-cpu==2.9.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
Werkzeug==3.0.3
scikit-learn==0.24.2  # For classification_report and confusion_matrix
split-folders==0.4.3
matplotlib==3.4.3
seaborn==0.11.2
pandas==1.3.4